test: 테스트 단계별 분리
